# docs/swagger.yaml
definitions:
  dto.AddBookRequest:
    properties:
      book_id:
        type: string
    type: object
...
  dto.AuthorResponse:
    description: Данные об авторе книги
    properties:
      id:
        description: |-
          Уникальный идентификатор автора (UUID)
          Example: "b6d46cd4-e89b-12d3-a456-426614174111"
        type: string
      name:
        description: |-
          Имя автора (обязательное поле)
          Required: true
          Example: "Дж. К. Роулинг"
        type: string
    type: object
...
  dto.BookListResponse:
    description: Ответ API со списком книг
    properties:
      books:
        description: Массив книг
        items:
          $ref: '#/definitions/dto.BookResponse'
        type: array
      next_cursor:
        description: |-
          Следующий маркер для пагинации (если есть)
          Example: "123e4567-e89b-12d3-a456-426614174002"
        type: string
    type: object
  dto.BookResponse:
    description: Ответ API на создание/обновление/получение книги
    properties:
      authors:
        description: Авторы книги (массив объектов)
        items:
          $ref: '#/definitions/dto.AuthorResponse'
        type: array
      average_rating:
        description: |-
          Средний рейтинг книги (из 10)
          Example: 8.5
        type: number
      cover_image:
        description: |-
          Обложка книги (URL)
          Example: "/uploads/123e4567-e89b-12d3-a456-426614174000.jpg"
        type: string
      description:
        description: |-
          Описание книги
          Example: "Первая книга о приключениях Гарри Поттера"
        type: string
      id:
        description: |-
          Уникальный идентификатор книги (UUID)
          Example: "123e4567-e89b-12d3-a456-426614174000"
        type: string
      title:
        description: |-
          Название книги (обязательное поле)
          Required: true
          Example: "Гарри Поттер и философский камень"
        type: string
    type: object
  dto.CreateBookRequest:
    properties:
      author_ids:
        items:
          type: string
        type: array
      cover_image:
        type: string
      description:
        type: string
      title:
        type: string
    type: object
...
  dto.UpdateBookRequest:
    description: запрос API на обновление книги
    properties:
      author_ids:
        items:
          type: string
        type: array
      cover_image:
        type: string
      description:
        type: string
      title:
        type: string
    type: object
...
  dto.UserBookResponse:
    properties:
      book_id:
        type: string
      created_at:
        type: string
      pages_read:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /books:
    get:
      description: Получить поджинированный список с курсором
      parameters:
      - description: UUID последней книги (для пагинации)
        in: query
        name: after_id
        type: string
      - description: Количество книг на страницу (по умолчанию 10)
        in: query
        name: limit
        type: integer
      - description: Данные для обновления книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.BookListResponse'
            type: array
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: маркерная пагинация книг
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Добавляет новую книгу в базу
      parameters:
      - description: Данные для создания книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BookResponse'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать книгу
      tags:
      - Books
  /books/{bookID}:
    get:
      description: получает книгу из базы по id
      parameters:
      - description: UUID книги
        in: path
        name: bookID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookResponse'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: получить книгу по ID
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: обновить книгу в базе по id
      parameters:
      - description: UUID книги
        in: path
        name: bookID
        required: true
        type: string
      - description: Данные для обновления книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.BookResponse'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: обновить книгу по ID
      tags:
      - Books
  /books/{bookID}/confirm:
    put:
      description: Модератор или админ может подтвердить книгу перед публикацией
      parameters:
      - description: UUID книги
        in: path
        name: bookID
        required: true
        type: string
      responses:
        "200":
          description: Book confirmed
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Подтвердить книгу
      tags:
      - Books
  /books/{bookID}/upload:
    post:
      consumes:
      - multipart/form-data
      description: Загрузка файла с обложкой книги (JPG, PNG)
      parameters:
      - description: UUID книги
        in: path
        name: bookID
        required: true
        type: string
      - description: Файл изображения (JPG, PNG, макс. 50 KB)
        in: formData
        name: cover
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid file format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Загрузить обложку книги
      tags:
      - Books
...
  /users/me/books:
    get:
      description: Возвращает список книг, добавленных пользователем
      produces:
      - application/json
      responses:
        "200":
          description: Список книг пользователя
          schema:
            items:
              $ref: '#/definitions/dto.UserBookResponse'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить список книг пользователя
      tags:
      - UserBooks
    post:
      consumes:
      - application/json
      description: Добавляет указанную книгу в список пользователя
      parameters:
      - description: ID книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.AddBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Книга добавлена в список'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить книгу в список пользователя
      tags:
      - UserBooks
  /users/me/books/{bookID}:
    delete:
      description: Удаляет указанную книгу из списка пользователя
      parameters:
      - description: UUID книги
        in: path
        name: bookID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Книга удалена из списка'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить книгу из списка пользователя
      tags:
      - UserBooks
...
swagger: "2.0"